backend/requirements.txt
psycopg2-binary
python-dotenv
fastapi
uvicorn
python-jose
passlib[bcrypt]
pymongo
dnspython